+++ src/imgui.lua
@@ -1,6 +1,11 @@
 ---@meta _
 ---@diagnostic disable
 
+local previousConfig = {
+    dress = nil,
+    random_each_run = nil,
+}
+
 rom.gui.add_imgui(function()
     if rom.ImGui.Begin("Dress Selector") then
         drawMenu()
@@ -20,12 +25,15 @@ function drawMenu()
     if rom.ImGui.BeginCombo("###dress", config.dress) then
         for _, dressName in ipairs(mod.DressDisplayOrder) do
             if rom.ImGui.Selectable(dressName, (dressName == config.dress)) then
-                local  dressGrannyTexture = mod.GetDressGrannyTexture(dressName)
-                config.dress = dressName
-                config.random_each_run = false
-                mod.UpdateSkin(dressGrannyTexture)
+                if dressName ~= previousConfig then
+                    local  dressGrannyTexture = mod.GetDressGrannyTexture(dressName)
+                    config.dress = dressName
+                    previousConfig.dress = dressName
+                    config.random_each_run = false
+                    mod.UpdateSkin(dressGrannyTexture)
+                end
+                rom.ImGui.SetItemDefaultFocus()
             end
-            rom.ImGui.SetItemDefaultFocus()
         end
         rom.ImGui.EndCombo()
     end
@@ -33,10 +41,10 @@ function drawMenu()
     rom.ImGui.Separator()
 
     local value, checked = rom.ImGui.Checkbox("Random Dress Each Run", config.random_each_run)
-    if checked then
+    if checked and value ~= previousConfig.random_each_run then
         config.random_each_run = value
+        previousConfig.random_each_run = value
         if value then
-            
             mod.UpdateSkin(mod.GetDressGrannyTexture(mod.GetCurrentRunDress()))
         else
             -- mod.ClearRunDressData()